refactor(analytics): move ViewTrafficSource out of the service layer
The featured tab needs the traffic-source enum at its navigation call site,
and the UI/service boundary ratchet correctly rejected importing a service
from a screen. The enum and its tag mapping are a domain value type rather
than service behavior, so they move to lib/models and the publisher
re-exports them — every existing caller keeps its current import.
